Move `SITEURL` variable to publishconf.py
According to a [comment by the
developer](https://stackoverflow.com/a/20845195), `pelicanconf.py`
should be thought of as the local settings and `publishconf.py` should
be thought of as the deployment settings. I typically end up commenting
out the `SITEURL` variable whenever I run `make html` to view a local
copy of the site, so it makes sense just to leave it out completely.
Furthermore, I noticed that the `SITEURL` in `publishconf.py` was set to
http instead of https, so I switched to https.