Add rst file to wrap text of CV
This rst file uses the `include` directive to wrap the full text of my
CV. This file is symlinked from the cv submodule directory to this
directory. In this way, I can modify the CV repo, pull the changes into
the submodule, and rebuild this site without having to change cv.rst.